Refactor toolbar into a separate file, add support for gridview/scrollview toolbar, add support for toolbar above/below editor, improve height adjusting to accomodate new changes, and expose execCommand to everyone (for future use)
