+++ lib/src/html_editor_mobile.dart
@@ -1,4 +1,5 @@
-import 'package:html_editor_enhanced/html_editor.dart';
+import 'package:html_editor_enhanced/html_editor.dart' hide HtmlEditorController;
+import 'package:html_editor_enhanced/src/html_editor_controller_mobile.dart';
 import 'package:html_editor_enhanced/src/widgets/html_editor_widget_mobile.dart'
     as html;
 import 'package:flutter/foundation.dart';
